update folder estructure
