+++ lib/domain/core/advetisement_model.dart
@@ -1,17 +1,15 @@
-import 'package:streaming_front_app/features/user_related/domain/entities/advertisement_entity.dart';
-
-class AdvertisementModel extends AdvertisementEntity {
-  const AdvertisementModel({
+class AdvertisementEntity {
+  const AdvertisementEntity({
     String? id,
     List<int>? image,
   });
 
-  factory AdvertisementModel.fromJson(Map<String, dynamic> json) {
+  factory AdvertisementEntity.fromJson(Map<String, dynamic> json) {
     // elements needed to get an image as buffer
     List<dynamic> bufferDynamic = json["image"]["data"];
     List<int> bufferInt = bufferDynamic.map((e) => e as int).toList();
 
-    return AdvertisementModel(
+    return AdvertisementEntity(
       id: json['id'] as String, //?? "",
       image: bufferInt,
     );
