--- lib/features/user_related/domain/entities/advertisement_entity.dart
@@ -1,16 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-class AdvertisementEntity extends Equatable {
-  final String? id;
-  final List<int>? image;
-
-  const AdvertisementEntity({this.id, this.image});
-
-  @override
-  List<Object?> get props {
-    return [
-      id,
-      image,
-    ];
-  }
-}
