+++ lib/presentation/pages/multimedia_related/artist_view.dart
