+++ lib/presentation/pages/multimedia_related/widgets/artist_cover_widget.dart
