+++ lib/presentation/pages/multimedia_related/widgets/simple_track_list_element.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:streaming_front_app/features/multimedia_related/domain/entities/track.dart';
+import 'package:streaming_front_app/domain/multimedia_related/track.dart';
 
 class SimpleTrackListElement extends StatelessWidget {
   const SimpleTrackListElement(this.trackToDisplay, {super.key});
