+++ lib/presentation/pages/user_related/iniciar_sesion.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:streaming_front_app/features/user_related/presentation/widgets/button_widget.dart';
-import 'package:streaming_front_app/features/user_related/presentation/widgets/default_background.dart';
-import 'package:streaming_front_app/features/user_related/presentation/widgets/generic_TextFormField.dart';
+
+import '../core/widgets/button_widget.dart';
+import '../core/widgets/default_background.dart';
+import '../core/widgets/generic_TextFormField.dart';
 
 class NewSesion extends StatefulWidget {
   const NewSesion({super.key, required this.title});
